pyproject.toml
[tool.poetry]
name = "command-line-loom"
version = "0.0.3"
description = ""
authors = ["fergus <fergusfettes@gmail.com>"]
homepage = "https://github.com/fergusfettes/command-line-loom"
readme = "README.md"
packages = [{include = "cll"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
pyyaml = "^6.0"
openai = "^0.27.0"
click = "^8.1.3"
tiktoken = "^0.3.0"
turbo-text-transformer-prompts = "^0.1.10"
python-dotenv = "^1.0.0"
click-shell = "^2.1"
typer = "^0.9.0"
rich = "^13.3.5"
iterfzf = "^0.5.0.20.0"
llama-index = "^0.6.8"
typer-shell = "^0.1.4"
networkx = "^3.1"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.11"
ipython = "^8.10.0"
memory-profiler = "^0.61.0"

[tool.poetry.scripts]
cll = "cll.__main__:main"

[project]
name = "command-line-loom"
version = "0.2.1"
dynamic = ["dependencies"]

[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}

[project.scripts]
cll = "cll.__main__:main"

[tool.setuptools]
packages = ["cll"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
